+++ src/main/java/com/jrpbjr/agendamentofinanceiro/payload/service/OperacaoService.java
@@ -2,10 +2,12 @@ package com.jrpbjr.agendamentofinanceiro.payload.service;
 import com.jrpbjr.agendamentofinanceiro.payload.exception.NegocioException;
 import com.jrpbjr.agendamentofinanceiro.payload.model.OperacaoModel;
 
+import java.math.BigDecimal;
+
 
 public interface  OperacaoService {
 
-    public Double calcularTaxa(OperacaoModel operacaoModel);
+    public BigDecimal calcularTaxa(OperacaoModel operacaoModel);
 
     public void validarTaxa(OperacaoModel operacaoModel)
         throws NegocioException;
